+++ rust/src/consolidate/otl/gsub_ligature.rs
@@ -75,5 +75,5 @@ pub unsafe fn consolidate_gsub_ligature(
         k = k.wrapping_add(1);
     }
     subtable_gsub_ligature_replace(subtable, nt);
-    return (*subtable).len() == 0 as usize;
+    return (*subtable).len() == 0_usize;
 }
