+++ rust/src/table/otl/subtables/gpos_cursive.rs
@@ -124,7 +124,7 @@ pub unsafe fn otl_gpos_dump_cursive(mut _subtable: *const Subtable) -> *mut Buil
     let st: *mut BuiltValue = json_object_new((*subtable).len());
     let mut j: GlyphId = 0 as GlyphId;
     while (j as usize) < (*subtable).len() {
-        let rec: *mut BuiltValue = json_object_new(2 as usize);
+        let rec: *mut BuiltValue = json_object_new(2_usize);
         json_object_push(
             rec,
             b"enter\0" as *const u8 as *const ::core::ffi::c_char,
@@ -191,7 +191,7 @@ pub unsafe fn otfcc_build_gpos_cursive(
         j = j.wrapping_add(1);
     }
     let root: *mut BkBlock = bk_new_block(&[
-        bk_int(BkCellType::B16, 1 as u32),
+        bk_int(BkCellType::B16, 1_u32),
         bk_ptr(
             BkCellType::P16,
             bk_new_block_from_buffer(build_coverage(cov)),
